refactor : 유저 예외 패키지 생성 및 예외 위치 수정
