+++ board/board/src/main/java/com/ys/board/domain/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.ys.board.domain.user.service;
 
 import com.ys.board.common.exception.EntityNotFoundException;
-import com.ys.board.domain.user.DuplicateNameException;
+import com.ys.board.domain.user.exception.DuplicateNameException;
 import com.ys.board.domain.user.model.User;
 import com.ys.board.domain.user.api.UserCreateRequest;
 import com.ys.board.domain.user.api.UserCreateResponse;
